Sistemazioni sintattiche di alcune classi
- Modifica del nome del package "DAO" in "dao";
- Cambiata convenzione sui prezzi di un immobile mutandola da double ad int;
- Accorgimenti tecnici su alcune classi al fine non visualizzare errori di compatibilità fra metodi;
- Modifica provvisoria della view di avvio nella classe main

+++ src/main/java/gui/ViewCaricaImmobile.java
@@ -235,9 +235,9 @@ public class ViewCaricaImmobile extends JFrame {
 		                return;
 		            }
 
-		            double prezzo;
+		            int prezzo;
 		            try {
-		                prezzo = Double.parseDouble(prezzofield.getText().trim());
+		                prezzo = Integer.parseInt(prezzofield.getText().trim());
 		            } catch (NumberFormatException ex) {
 		                JOptionPane.showMessageDialog(null, "Inserisci un numero valido per il prezzo");
 		                return;
